Renaming zero order unit model files to allow for single inlet single outlet case
